// src/types/tyoe.ts
export interface Post {
    userId: number;
    id: number;
    title: string;
    body: string;
}


export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  phone: string;
  website: string;
  company: { name: string; catchPhrase: string; bs: string };
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
  };
}